requirements.txt
SpeechRecognition
pyaudio
arabic-reshaper
python-bidi
PyArabic
nltk
webrtcvad
